Add hparams and job status support to SummaryWriter
Introduced new modules `hparams` and `status` to support hyperparameter configurations and job status reporting in TensorBoard. Extended `SummaryWriter` with methods to add hyperparameter configs, scalar metrics, and job statuses. Updated dependencies and added a new example demonstrating these capabilities.

+++ tensorboard-rs/src/record_writer.rs
@@ -12,8 +12,8 @@ impl<W: Write> RecordWriter<W> {
     }
     pub fn write(&mut self, data: &[u8]) -> std::io::Result<()>{
         let header = data.len() as u64;
-        let header_crc = (masked_crc32c(&(header.to_le_bytes())) as u32).to_le_bytes();
-        let footer_crc = (masked_crc32c(data) as u32).to_le_bytes();
+        let header_crc = masked_crc32c(&(header.to_le_bytes())).to_le_bytes();
+        let footer_crc = masked_crc32c(data).to_le_bytes();
         let header = header.to_le_bytes();
 
         self._writer.write_all(&header)?;
